third_party/blink/web_tests/external/wpt/web-animations/responsive/lineHeight-expected.txt

This is a testharness.js-based test.
[PASS] lineHeight responsive to style changes
[FAIL] lineHeight clamped to 0px on keyframes
  assert_equals: expected "20px" but got "0px"
[PASS] lineHeight responsive to inherited changes from keyword
[PASS] lineHeight responsive to inherited changes from number
[PASS] lineHeight responsive to inherited changes from length
Harness: the test ran to completion.
